backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
requests
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
